pyproject.toml
[tool.poetry]
name = "svt"
version = "0.1.0"
description = ""
authors = ["mukappalambda <maokailan24@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
svt-cli = "svt.cli:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
numpy = "^1.24.2"
scipy = "^1.10.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
